refactor: split spawn state module
