+++ world.py
@@ -12,6 +12,7 @@ import settings as _cfg
 
 __path__ = [os.path.join(os.path.dirname(__file__), "world")]
 player = importlib.import_module(__name__ + ".player")
+spawns = importlib.import_module(__name__ + ".spawns")
 
 from world.player import (
     BADGE_DEFINITIONS, DEFAULT_AVATAR, ITEM_GREAT_BALL, ITEM_POKE_BALL,
@@ -24,6 +25,13 @@ from world.player import (
     level_bounds, level_for_xp, new_uid, onboarding_needed, record_badge_progress,
     room_in_bag, save, set_password, spend, spend_coins, take_item, team_for,
     type_badges_from_game_master, use,
+)
+
+from world.spawns import (
+    BONUS_SPAWNS, DESPAWNED, FORT_MODIFIERS, SPAWNS, _MAX_SPAWNS,
+    add_bonus_spawn, add_fort_modifier, bonus_spawns, drop_bonus_spawn,
+    fort_modifier, get_spawn, is_despawned, lured_forts, mark_despawned,
+    remember_spawn, remove_spawn,
 )
 
 HERE = player.HERE
@@ -457,41 +465,6 @@ def xp_multiplier():
     return 2 if item_active(301) else 1
 
 
-def add_fort_modifier(fort_id, item_id, minutes, by):
-    """Attach a Lure to a PokeStop."""
-    now = int(time.time() * 1000)
-    with _lock:
-        cur = FORT_MODIFIERS.get(fort_id)
-        if cur and cur.get("expires_ms", 0) > now:
-            return 2, None
-    if not take_item(int(item_id), 1):
-        return 4, None
-    mod = {"item": int(item_id), "expires_ms": now + int(minutes * 60000), "by": by}
-    with _lock:
-        FORT_MODIFIERS[fort_id] = mod
-    return 1, dict(mod)
-
-
-def fort_modifier(fort_id):
-    now = int(time.time() * 1000)
-    with _lock:
-        m = FORT_MODIFIERS.get(fort_id)
-        if not m:
-            return None
-        if m.get("expires_ms", 0) <= now:
-            FORT_MODIFIERS.pop(fort_id, None)
-            return None
-        return dict(m)
-
-
-def lured_forts():
-    now = int(time.time() * 1000)
-    with _lock:
-        for fid in [
-            k for k, m in FORT_MODIFIERS.items() if m.get("expires_ms", 0) <= now
-        ]:
-            FORT_MODIFIERS.pop(fid, None)
-        return {k: dict(v) for k, v in FORT_MODIFIERS.items()}
 
 
 def use_berry(encounter_id, mult):
@@ -513,58 +486,7 @@ def berry_mult(encounter_id, consume=False):
 
 GYMS = {}                          # fort_id -> [{uid, pokemon_id, cp, trainer, team}]
 BATTLES = {}                       # battle_id -> live battle state
-SPAWNS = {}                        # transient
-DESPAWNED = {}                     # encounter_id -> expiry_ms
-FORT_MODIFIERS = {}                # fort_id -> {item, expires_ms, by}
 RAID = {"on": False, "pokemon_id": 150, "cp": 3000, "trainer": "raid"}
-BONUS_SPAWNS = {}                  # username -> [ {eid,pid,cp,lat,lng,expires_ms} ]
-_MAX_SPAWNS = 4000
-
-
-def remember_spawn(encounter_id, pokemon_id, lat, lng, cp, spawn_id, expires_ms):
-    with _lock:
-        if len(SPAWNS) >= _MAX_SPAWNS:
-            for k in sorted(SPAWNS, key=lambda k: SPAWNS[k]["expires_ms"])[: _MAX_SPAWNS // 2]:
-                SPAWNS.pop(k, None)
-        SPAWNS[encounter_id] = {
-            "pokemon_id": pokemon_id,
-            "lat": lat,
-            "lng": lng,
-            "cp": cp,
-            "spawn_id": spawn_id,
-            "expires_ms": expires_ms,
-        }
-
-
-def get_spawn(encounter_id):
-    with _lock:
-        s = SPAWNS.get(encounter_id)
-        return dict(s) if s else None
-
-
-def remove_spawn(encounter_id):
-    with _lock:
-        SPAWNS.pop(encounter_id, None)
-
-
-def mark_despawned(encounter_id, until_ms):
-    with _lock:
-        DESPAWNED[encounter_id] = until_ms
-        if len(DESPAWNED) > 5000:
-            now = int(time.time() * 1000)
-            for k in [k for k, v in DESPAWNED.items() if v < now]:
-                DESPAWNED.pop(k, None)
-
-
-def is_despawned(encounter_id):
-    with _lock:
-        exp = DESPAWNED.get(encounter_id)
-        if exp is None:
-            return False
-        if exp < int(time.time() * 1000):
-            DESPAWNED.pop(encounter_id, None)
-            return False
-        return True
 
 
 # ==============================================================================
@@ -641,35 +563,6 @@ def is_raid_uid(fort_id, uid):
     return RAID["on"] and _raid_member(fort_id)["uid"] == uid
 
 
-def add_bonus_spawn(username, eid, pid, cp, lat, lng, expires_ms):
-    """A one-off wild Pokemon placed for ONE trainer."""
-    with _lock:
-        lst = BONUS_SPAWNS.setdefault(username, [])
-        lst.append(
-            {
-                "eid": eid,
-                "pid": pid,
-                "cp": cp,
-                "lat": lat,
-                "lng": lng,
-                "expires_ms": expires_ms,
-            }
-        )
-        del lst[:-10]
-
-
-def bonus_spawns(username):
-    now = int(time.time() * 1000)
-    with _lock:
-        lst = [b for b in BONUS_SPAWNS.get(username, []) if b["expires_ms"] > now]
-        BONUS_SPAWNS[username] = lst
-        return [dict(b) for b in lst]
-
-
-def drop_bonus_spawn(username, eid):
-    with _lock:
-        lst = BONUS_SPAWNS.get(username, [])
-        BONUS_SPAWNS[username] = [b for b in lst if b["eid"] != eid]
 
 
 # ==============================================================================
